+++ packages/js/client/src/__tests__/core/plugin-wrapper.spec.ts
@@ -1,59 +1,30 @@
 import {
-  initTestEnvironment,
-  stopTestEnvironment,
-} from "@web3api/test-env-js";
-import {
+  Web3ApiClientConfig,
   Plugin,
+  Web3ApiClient,
+  createWeb3ApiClient,
   PluginModule,
   PluginModules,
-} from "@web3api/core-js";
-import {
-  Web3ApiClientConfig,
-  Web3ApiClient,
 } from "../..";
-import { createWeb3ApiClient } from "../../createWeb3ApiClient";
 
 jest.setTimeout(200000);
 
+const defaultPlugins = [
+  "w3://ens/ipfs.web3api.eth",
+  "w3://ens/ens.web3api.eth",
+  "w3://ens/ethereum.web3api.eth",
+  "w3://ens/http.web3api.eth",
+  "w3://ens/js-logger.web3api.eth",
+  "w3://ens/uts46.web3api.eth",
+  "w3://ens/sha3.web3api.eth",
+  "w3://ens/graph-node.web3api.eth",
+  "w3://ens/fs.web3api.eth",
+];
+
 describe("plugin-wrapper", () => {
-  let ipfsProvider: string;
-  let ethProvider: string;
-  let ensAddress: string;
-
-  beforeAll(async () => {
-    const {
-      ipfs,
-      ethereum,
-      ensAddress: ens,
-    } = await initTestEnvironment();
-    ipfsProvider = ipfs;
-    ethProvider = ethereum;
-    ensAddress = ens;
-  });
-
-  afterAll(async () => {
-    await stopTestEnvironment();
-  });
-
   const getClient = async (config?: Partial<Web3ApiClientConfig>) => {
     const client = await createWeb3ApiClient(
-      {
-        ethereum: {
-          networks: {
-            testnet: {
-              provider: ethProvider,
-            },
-          },
-        },
-        ipfs: { provider: ipfsProvider },
-        ens: {
-          query: {
-            addresses: {
-              testnet: ensAddress,
-            },
-          },
-        },
-      },
+      {},
       config
     );
 
@@ -66,13 +37,13 @@ describe("plugin-wrapper", () => {
     }
 
     class Query extends PluginModule<Config> {
-      async getMap(_: unknown) { return this.config.map }
+      async getMap(_: unknown) {
+        return this.config.map;
+      }
     }
 
     class Mutation extends PluginModule<Config> {
-      updateMap(input: {
-        map: Map<string, number>;
-      }): Map<string, number> {
+      updateMap(input: { map: Map<string, number> }): Map<string, number> {
         for (const key of input.map.keys()) {
           this.config.map.set(
             key,
@@ -84,8 +55,7 @@ describe("plugin-wrapper", () => {
     }
 
     class MockMapPlugin implements Plugin {
-
-      private map = new Map().set("a", 1).set("b", 2)
+      private map = new Map().set("a", 1).set("b", 2);
 
       getModules(): PluginModules {
         return {
@@ -106,17 +76,6 @@ describe("plugin-wrapper", () => {
 
   test("plugin registration - with default plugins", () => {
     const implementationUri = "w3://ens/some-implementation.eth";
-    const defaultPlugins = [
-      "w3://ens/ipfs.web3api.eth",
-      "w3://ens/ens.web3api.eth",
-      "w3://ens/ethereum.web3api.eth",
-      "w3://ens/http.web3api.eth",
-      "w3://ens/js-logger.web3api.eth",
-      "w3://ens/uts46.web3api.eth",
-      "w3://ens/sha3.web3api.eth",
-      "w3://ens/graph-node.web3api.eth",
-      "w3://ens/fs.web3api.eth",
-    ];
 
     const client = new Web3ApiClient({
       plugins: [
@@ -143,7 +102,7 @@ describe("plugin-wrapper", () => {
     const schema: string = await client.getSchema(
       "w3://ens/js-logger.web3api.eth"
     );
-    
+
     expect(schema).toStrictEqual(
       `### Web3API Header START ###
 scalar UInt
@@ -271,5 +230,79 @@ enum Logger_LogLevel @imported(
     expect(mutationEnv.data?.updateMap).toMatchObject(
       new Map<string, number>().set("a", 1).set("b", 3).set("c", 5)
     );
+  });
+
+  test("plugin registration - with plugin override", async () => {
+    const pluginUriToOverride = defaultPlugins[0];
+
+    const pluginPackage = {
+      factory: () => ({} as Plugin),
+      manifest: {
+        schema: "",
+        implements: [],
+      },
+    };
+
+    const client = new Web3ApiClient({
+      plugins: [
+        {
+          uri: pluginUriToOverride,
+          plugin: pluginPackage,
+        },
+      ],
+    });
+
+    const pluginUris = client.getPlugins().map((x) => x.uri.uri);
+
+    expect(pluginUris).toEqual(defaultPlugins);
+
+    const registeredPlugin = client
+      .getPlugins()
+      .find((x) => x.uri.uri === pluginUriToOverride);
+
+    expect(registeredPlugin?.plugin).toEqual(pluginPackage);
+  });
+
+  test("plugin registration - with multiple plugin overrides", async () => {
+    const pluginUriToOverride = defaultPlugins[0];
+
+    const pluginPackage1 = {
+      factory: () => ({} as Plugin),
+      manifest: {
+        schema: "",
+        implements: [],
+      },
+    };
+
+    const pluginPackage2 = {
+      factory: () => ({} as Plugin),
+      manifest: {
+        schema: "",
+        implements: [],
+      },
+    };
+
+    const client = new Web3ApiClient({
+      plugins: [
+        {
+          uri: pluginUriToOverride,
+          plugin: pluginPackage1,
+        },
+        {
+          uri: pluginUriToOverride,
+          plugin: pluginPackage2,
+        },
+      ],
+    });
+
+    const pluginUris = client.getPlugins().map((x) => x.uri.uri);
+
+    expect(pluginUris).toEqual(defaultPlugins);
+
+    const registeredPlugin = client
+      .getPlugins()
+      .find((x) => x.uri.uri === pluginUriToOverride);
+
+    expect(registeredPlugin?.plugin).toEqual(pluginPackage1);
   });
 });