+++ packages/wasm/rs/src/msgpack/mod.rs
@@ -1,4 +1,3 @@
-pub mod context;
 pub mod data_view;
 pub mod error;
 pub mod format;
@@ -8,7 +7,6 @@ pub mod read_decoder;
 pub mod write;
 pub mod write_encoder;
 
-pub use context::Context;
 pub use data_view::DataView;
 pub use error::{DecodeError, EncodeError, EnumTypeError};
 pub use format::Format;